+++ NEWS.md
@@ -1,7 +1,7 @@
-# modelsummary 0.9.6.9000
+# modelsummary 0.10.0
 
 * New `exponentiate` argument for `modelsummary()` and `modelplot()`
-* `gof_map` accepts a vector like: `c("rmse", "nobs", "r.squared")`
+* `gof_map` accepts a vector such as `c("rmse", "nobs", "r.squared")`
 * Drop `rlang` dependency
 * Bug fixes